requires admin role to update and delete users

+++ StudyMate/app/Http/Controllers/Admin/UsersController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
+use Gate;
 use Illuminate\Http\Request;
 
 class UsersController extends Controller
@@ -33,6 +34,9 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        if(Gate::denies('edit-users')){
+            return redirect(route('admin.users.index'));
+        }
         $roles = Role::all();
 
         return view('admin.users.edit')->with([
@@ -63,6 +67,9 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if(Gate::denies('destroy-users')){
+            return redirect(route('admin.users.index'));
+        }
         $user->roles()->detach();
         $user->delete();
